+++ src/Model/Config/PluginConfig.php
@@ -13,6 +13,9 @@ use srag\Plugins\Opencast\Model\Publication\Config\PublicationUsageGroup;
 use srag\Plugins\Opencast\Model\TermsOfUse\ToUManager;
 use srag\Plugins\Opencast\Model\User\xoctUser;
 use srag\Plugins\Opencast\Model\WorkflowParameter\Config\WorkflowParameter;
+use srag\Plugins\Opencast\Model\Metadata\Config\Event\MDFieldConfigEventAR;
+use srag\Plugins\Opencast\Model\Metadata\Config\Series\MDFieldConfigSeriesAR;
+use srag\Plugins\Opencast\Model\Workflow\WorkflowAR;
 use xoctCurl;
 use xoctCurlSettings;
 use xoctLog;
@@ -79,6 +82,7 @@ class PluginConfig extends ActiveRecord
     public const F_START_X_MINUTES_BEFORE_LIVE = 'start_x_minutes_before_live';
     public const F_PRESENTATION_NODE = 'presentation_node';
     public const F_LIVESTREAM_TYPE = 'livestream_type';
+    public const F_LIVESTREAM_BUFFERED = 'livestream_buffered';
     public const F_ENABLE_CHAT = 'enable_chat';
 
     public const F_REPORT_QUALITY = 'report_quality';
@@ -122,7 +126,9 @@ class PluginConfig extends ActiveRecord
     public const PAELLA_LANG_PATH = 'Customizing/global/plugins/Services/Repository/RepositoryObject/OpenCast/js/opencast/src/Paella/lang';
     public const PAELLA_DEFAULT_THEME = 'Customizing/global/plugins/Services/Repository/RepositoryObject/OpenCast/js/opencast/src/Paella/default_theme/opencast_theme.json';
     public const PAELLA_DEFAULT_THEME_LIVE = 'Customizing/global/plugins/Services/Repository/RepositoryObject/OpenCast/js/opencast/src/Paella/default_theme/opencast_live_theme.json';
+    public const PAELLA_DEFAULT_THEME_LIVE_BUFFERED = 'Customizing/global/plugins/Services/Repository/RepositoryObject/OpenCast/js/opencast/src/Paella/default_theme/opencast_live_buffered_theme.json';
 
+    public const F_WORKFLOWS_TAGS = 'config_workflows_tags';
     public const F_PAELLA_PREVIEW_FALLBACK = 'paella_config_preview_fallback';
     public const F_PAELLA_PREVIEW_FALLBACK_URL = 'paella_config_preview_fallback_url';
     public const PAELLA_DEFAULT_PREVIEW = 'Customizing/global/plugins/Services/Repository/RepositoryObject/OpenCast/templates/images/default_preview.png';
@@ -198,6 +204,54 @@ class PluginConfig extends ActiveRecord
         }
 
         /**
+         * @var $xoctMDFieldConfigEventAR MDFieldConfigEventAR
+         */
+        $xoct_md_field_event = $domxml->getElementsByTagName('xoct_md_field_event');
+
+        // Clear MDFieldConfigEventAR
+        MDFieldConfigEventAR::flushDB();
+
+        foreach ($xoct_md_field_event as $node) {
+            $xoctMDFieldConfigEventAR = new MDFieldConfigEventAR();
+            $xoctMDFieldConfigEventAR->setSort($node->getElementsByTagName('sort')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setFieldId($node->getElementsByTagName('field_id')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setTitleDe($node->getElementsByTagName('title_de')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setTitleEn($node->getElementsByTagName('title_en')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setVisibleForPermissions(
+                $node->getElementsByTagName('visible_for_permissions')->item(0)->nodeValue
+            );
+            $xoctMDFieldConfigEventAR->setPrefill($node->getElementsByTagName('prefill')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setReadOnly($node->getElementsByTagName('read_only')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setRequired($node->getElementsByTagName('required')->item(0)->nodeValue);
+            $xoctMDFieldConfigEventAR->setValuesFromEditableString($node->getElementsByTagName('values')->item(0)->nodeValue ?? '');
+            $xoctMDFieldConfigEventAR->store();
+        }
+
+        /**
+         * @var $xoctMDFieldConfigSeriesAR MDFieldConfigSeriesAR
+         */
+        $xoct_md_field_series = $domxml->getElementsByTagName('xoct_md_field_series');
+
+        // Clear MDFieldConfigSeriesAR
+        MDFieldConfigSeriesAR::flushDB();
+
+        foreach ($xoct_md_field_series as $node) {
+            $xoctMDFieldConfigSeriesAR = new MDFieldConfigSeriesAR();
+            $xoctMDFieldConfigSeriesAR->setSort($node->getElementsByTagName('sort')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setFieldId($node->getElementsByTagName('field_id')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setTitleDe($node->getElementsByTagName('title_de')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setTitleEn($node->getElementsByTagName('title_en')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setVisibleForPermissions(
+                $node->getElementsByTagName('visible_for_permissions')->item(0)->nodeValue
+            );
+            $xoctMDFieldConfigSeriesAR->setPrefill($node->getElementsByTagName('prefill')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setReadOnly($node->getElementsByTagName('read_only')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setRequired($node->getElementsByTagName('required')->item(0)->nodeValue);
+            $xoctMDFieldConfigSeriesAR->setValuesFromEditableString($node->getElementsByTagName('values')->item(0)->nodeValue ?? '');
+            $xoctMDFieldConfigSeriesAR->store();
+        }
+
+        /**
          * @var $xoctWorkflowParameter WorkflowParameter
          */
         $xoct_workflow_parameter = $domxml->getElementsByTagName('xoct_workflow_parameter');
@@ -222,6 +276,24 @@ class PluginConfig extends ActiveRecord
             } else {
                 $xoctWorkflowParameter->update();
             }
+        }
+
+        /**
+         * @var $xoctWorkflowParameter WorkflowParameter
+         */
+        $xoct_workflow = $domxml->getElementsByTagName('xoct_workflow');
+
+        // We need to reset the workflow table.
+        WorkflowAR::flushDB();
+
+        foreach ($xoct_workflow as $node) {
+            $xoctWorkflow = new WorkflowAR();
+            $xoctWorkflow->setWorkflowId($node->getElementsByTagName('workflow_id')->item(0)->nodeValue);
+            $xoctWorkflow->setTitle($node->getElementsByTagName('title')->item(0)->nodeValue ?? '');
+            $xoctWorkflow->setDescription($node->getElementsByTagName('description')->item(0)->nodeValue ?? '');
+            $xoctWorkflow->setTags($node->getElementsByTagName('tags')->item(0)->nodeValue ?? '');
+            $xoctWorkflow->setConfigPanel($node->getElementsByTagName('config_panel')->item(0)->nodeValue ?? '');
+            $xoctWorkflow->create();
         }
 
         /**
@@ -354,6 +426,79 @@ class PluginConfig extends ActiveRecord
             $xml_xoctConf->appendChild(new DOMElement('value'))->appendChild(new DOMCdataSection($value));
         }
 
+
+        // xoctMDFieldConfigEventARs
+        $xml_xoctMDFieldConfigEventARs = $config->appendChild(new DOMElement('xoct_md_field_events'));
+        /**
+         * @var $xoctMDFieldConfigEventARs MDFieldConfigEventAR
+         */
+        foreach (MDFieldConfigEventAR::get() as $xoctMDFieldConfigEventAR) {
+            $xml_xoctMDE = $xml_xoctMDFieldConfigEventARs->appendChild(new DOMElement('xoct_md_field_event'));
+            $xml_xoctMDE->appendChild(new DOMElement('sort'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getSort())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('field_id'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getFieldId())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('title_de'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getTitle('de'))
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('title_en'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getTitle('en'))
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('visible_for_permissions'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getVisibleForPermissions())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('prefill'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getPrefill())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('read_only'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->isReadOnly())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('required'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->isRequired())
+            );
+            $xml_xoctMDE->appendChild(new DOMElement('values'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigEventAR->getValuesAsEditableString())
+            );
+        }
+
+        // xoctMDFieldConfigSeriesARs
+        $xml_xoctMDFieldConfigSeriesARs = $config->appendChild(new DOMElement('xoct_md_field_serieses'));
+        /**
+         * @var $xoctMDFieldConfigSeriesARs MDFieldConfigSeriesAR
+         */
+        foreach (MDFieldConfigSeriesAR::get() as $xoctMDFieldConfigSeriesAR) {
+            $xml_xoctMDS = $xml_xoctMDFieldConfigSeriesARs->appendChild(new DOMElement('xoct_md_field_series'));
+            $xml_xoctMDS->appendChild(new DOMElement('sort'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getSort())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('field_id'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getFieldId())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('title_de'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getTitle('de'))
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('title_en'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getTitle('en'))
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('visible_for_permissions'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getVisibleForPermissions())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('prefill'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getPrefill())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('read_only'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->isReadOnly())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('required'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->isRequired())
+            );
+            $xml_xoctMDS->appendChild(new DOMElement('values'))->appendChild(
+                new DOMCdataSection($xoctMDFieldConfigSeriesAR->getValuesAsEditableString())
+            );
+        }
+
         // xoctWorkflowParameters
         $xml_xoctWorkflowParameters = $config->appendChild(new DOMElement('xoct_workflow_parameters'));
         /**
@@ -375,6 +520,30 @@ class PluginConfig extends ActiveRecord
             );
             $xml_xoctPU->appendChild(new DOMElement('default_value_admin'))->appendChild(
                 new DOMCdataSection($xoctWorkflowParameter->getDefaultValueAdmin())
+            );
+        }
+
+        // xoctWorkflows
+        $xml_xoctWorkflows = $config->appendChild(new DOMElement('xoct_workflows'));
+        /**
+         * @var $xoctWorkflowAR WorkflowAR
+         */
+        foreach (WorkflowAR::get() as $xoctWorkflows) {
+            $xml_xoctWf = $xml_xoctWorkflows->appendChild(new DOMElement('xoct_workflow'));
+            $xml_xoctWf->appendChild(new DOMElement('workflow_id'))->appendChild(
+                new DOMCdataSection($xoctWorkflows->getWorkflowId())
+            );
+            $xml_xoctWf->appendChild(new DOMElement('title'))->appendChild(
+                new DOMCdataSection($xoctWorkflows->getTitle() ?? '')
+            );
+            $xml_xoctWf->appendChild(new DOMElement('description'))->appendChild(
+                new DOMCdataSection($xoctWorkflows->getDescription() ?? '')
+            );
+            $xml_xoctWf->appendChild(new DOMElement('tags'))->appendChild(
+                new DOMCdataSection($xoctWorkflows->getTags() ?? '')
+            );
+            $xml_xoctWf->appendChild(new DOMElement('config_panel'))->appendChild(
+                new DOMCdataSection($xoctWorkflows->getConfigPanel() ?? '')
             );
         }